Add further design notes after consulting standard texts on design patterns
A balance is being sought between clean architecture and not over-designing
by including known principles for the sake of including principles. Rather
the approach is to build incrementally making use of patterns to allow
flexibility to make change easier in ways in which change is perceived to
be very likely to be needed.